Move UpdateTimestampListener into separate class
